+++ src/lexer/utils/index.ts
@@ -3,3 +3,4 @@ export * from "./string";
 export * from "./word";
 export * as javascript from "./javascript";
 export * as json from "./json";
+export * from "./numeric";